Refactor directories to follow duck pattern
