+++ src/App.js
@@ -1,13 +1,15 @@
 import React from 'react';
+// import { configureStore } from '@reduxjs/toolkit';
 import {
   Routes,
   Route,
 } from 'react-router-dom';
 import './App.css';
-import Books from './pages/Books';
+import Books from './redux/books/Books';
 import Navbar from './components/Navbar';
-import Categories from './pages/Categories';
+import Categories from './redux/categories/Categories';
 
+// const store = configureStore();
 function App() {
   return (
     <div className="App">
