+++ src/redux/categories/Categories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/Categories.css';
+import '../../styles/Categories.css';
 
 const Categories = () => (
   <div className="big-container">
